src/requirements.txt
Django==1.11.8
django-crispy-forms==1.7.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2018.3
